requirements.txt
TTS==0.21.1
elevenlabs==0.2.27
openai==1.12.0
python-dotenv==1.0.0
flask==3.0.2
gunicorn==21.2.0 